third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/ui-source-code-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests UISourceCode class.


Running: testUISourceCode
Content is requested from SourceCodeProvider.
Callback 1 is invoked.
Callback 2 is invoked.
Callback 3 is invoked.
